Move all CBMC stubs to a single folder (#918)
