+++ cbmc/proofs/IotMqtt_DeserializeSuback/_IotMqtt_DeserializeSuback_harness.c
@@ -1,70 +1,36 @@
+/*
+ * IoT MQTT V2.1.0
+ * Copyright (C) Amazon.com, Inc. or its affiliates.  All Rights Reserved.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+/**
+ * @file _IotMqtt_DeserializeSuback_harness.c
+ * @brief Implements the proof harness for _IotMqtt_DeserializeSuback function.
+ */
+
 #include "iot_config.h"
 #include "private/iot_mqtt_internal.h"
 
 #include <stdlib.h>
 
 #include "mqtt_state.h"
-
-/****************************************************************
- * Stub out RemoveAllMatches.
- *
- * The only invocation of RemoveAllMatches in DeserializeSuback is to
- * remove subscriptions from the connection's subscription list.  This
- * stub abstracts RemoveAllMatches by replacing the list with an
- * unconstrained list of subscriptions.  We don't even bother to fill
- * in the topic subscription filters.  By design, any actual use of
- * the list in subsequent code (there is none) will trigger pointer
- * errors.
- *****************************************************************/
-
-void IotListDouble_RemoveAllMatches( const IotListDouble_t * const pList,
-                                     bool ( *isMatch )( const IotLink_t * const pOperationLink,
-                                                        void * pCompare ),
-                                     void * pMatch,
-                                     void ( *freeElement )( void * pData ),
-                                     size_t linkOffset )
-{
-    IotListDouble_t * sentinel = pList->pNext->pPrevious;
-
-    size_t num_elts;
-
-    __CPROVER_assume( num_elts < SUBSCRIPTION_COUNT_MAX );
-
-    /* Allocate lists of length at most 3 */
-    __CPROVER_assert( SUBSCRIPTION_COUNT_MAX <= 4,
-                      "Subscription lists limited to 3 elements" );
-
-    if( 1 <= num_elts )
-    {
-        _mqttSubscription_t * pElt = malloc( sizeof( *pElt ) );
-        sentinel->pNext = &pElt->link;
-        pElt->link.pPrevious = sentinel;
-        pElt->link.pNext = sentinel;
-        sentinel->pPrevious = &pElt->link;
-    }
-
-    if( 2 <= num_elts )
-    {
-        _mqttSubscription_t * pElt = malloc( sizeof( *pElt ) );
-        sentinel->pNext = &pElt->link;
-        pElt->link.pPrevious = sentinel;
-        pElt->link.pNext = sentinel;
-        sentinel->pPrevious = &pElt->link;
-    }
-
-    if( 3 <= num_elts )
-    {
-        _mqttSubscription_t * pElt = malloc( sizeof( *pElt ) );
-        sentinel->pNext = &pElt->link;
-        pElt->link.pPrevious = sentinel;
-        pElt->link.pNext = sentinel;
-        sentinel->pPrevious = &pElt->link;
-    }
-}
-
-/****************************************************************
-* Proof harness
-****************************************************************/
 
 void harness()
 {
@@ -85,5 +51,3 @@ void harness()
     /* Argument must be a nonnull pointer */
     _IotMqtt_DeserializeSuback( &suback );
 }
-
-/****************************************************************/
