+++ cbmc/proofs/IotMqtt_ReceiveCallback/IotMqtt_ReceiveCallback_harness.c
@@ -36,201 +36,6 @@
 #include "mqtt_state.h"
 
 /****************************************************************
-* Type definitions used by the IoT List Double remove functions
-****************************************************************/
-
-typedef bool ( * MatchFunction_t )( const IotLink_t * const pOperationLink,
-                                    void * pCompare );
-typedef void ( * FreeElementFunction_t )( void * pData );
-
-/****************************************************************
-* We assume the IoT List Double remove functions are memory safe.
-*
-* We abstract the list remove functions for performance reasons.  Our
-* abstraction replaces the original list with an unconstrained list.
-* Our abstraction proves that none of the elements on the original
-* list are accessed after the remove: We free all elements on the
-* original list, so that any later access will be caught as a
-* use-after-free error.
-****************************************************************/
-
-void IotListDouble_RemoveAllMatches( const IotListDouble_t * const pList,
-                                     MatchFunction_t isMatch,
-                                     void * pMatch,
-                                     FreeElementFunction_t freeElement,
-                                     size_t linkOffset )
-{
-    free_IotMqttSubscriptionList( pList );
-    allocate_IotMqttSubscriptionList( pList, SUBSCRIPTION_COUNT_MAX - 1 );
-}
-
-/****************************************************************/
-
-void IotListDouble_RemoveAll( const IotListDouble_t * const pList,
-                              FreeElementFunction_t freeElement,
-                              size_t linkOffset )
-{
-    free_IotMqttSubscriptionList( pList );
-    allocate_IotMqttSubscriptionList( pList, SUBSCRIPTION_COUNT_MAX - 1 );
-}
-
-/****************************************************************
-* We assume the TaskPool functions are memory safe.
-*
-* We abstract the task pool functions (we assume they have no side
-* effects and return an unconstrained value), but there are some
-* assertions in the code that the task pool functions return a
-* constrained set of values.  We model these functions as simply
-* returning one of those values.
-****************************************************************/
-
-IotTaskPoolError_t IotTaskPool_CreateJob( const IotTaskPoolRoutine_t userCallback,
-                                          void * const pUserContext,
-                                          IotTaskPoolJob_t * const pJob )
-{
-    assert( userCallback != NULL );
-    assert( pUserContext != NULL );
-    assert( pJob != NULL );
-
-    /*
-     * *_IotMqtt_ScheduleOperation says creating a new job should
-     * never fail when parameters are valid.
-     */
-    return IOT_TASKPOOL_SUCCESS;
-}
-
-/****************************************************************/
-
-IotTaskPoolError_t IotTaskPool_ScheduleDeferred( IotTaskPool_t * const pTaskPool,
-                                                 IotTaskPoolJob_t * const pJob,
-                                                 uint32_t timeMs )
-{
-    /* assert(pTaskPool != NULL); */
-    /* assert(pJob != NULL); */
-
-    /*
-     * *_IotMqtt_ScheduleOperation says a newly created job should
-     * never be invalid or illegal.
-     */
-    IotTaskPoolError_t error;
-
-    __CPROVER_assume( error != IOT_TASKPOOL_BAD_PARAMETER );
-    __CPROVER_assume( error != IOT_TASKPOOL_ILLEGAL_OPERATION );
-    return error;
-}
-
-/****************************************************************
-* We abstract the Notify method by assuming it havocs subscription and
-* operation lists.
-*
-* These functions are used to havoc the lists.  The lists are written
-* and then never again read.  We model these updates by simply freeing
-* the lists, and trusting CBMC to flag a use-after-free error if an
-* element of the list is accessed.
-****************************************************************/
-
-void * invalid_pointer();
-
-IotListDouble_t * destroy_subscription_list_element( IotListDouble_t * pElt )
-{
-    if( pElt == NULL )
-    {
-        return NULL;
-    }
-
-    IotListDouble_t * pNext = pElt->pNext;
-    free( IotLink_Container( _mqttSubscription_t, pElt, link ) );
-    return pNext;
-}
-
-void destroy_subscription_list( IotListDouble_t * pList )
-{
-    if( pList == NULL )
-    {
-        return;
-    }
-
-    /* Assuming lists are of length at most 3 */
-
-    IotListDouble_t * pElt = pList->pNext;
-
-    if( pElt != pList )
-    {
-        pElt = destroy_subscription_list_element( pElt );
-    }
-
-    if( pElt != pList )
-    {
-        pElt = destroy_subscription_list_element( pElt );
-    }
-
-    if( pElt != pList )
-    {
-        pElt = destroy_subscription_list_element( pElt );
-    }
-
-    pList->pNext = invalid_pointer();
-    pList->pPrevious = invalid_pointer();
-}
-
-IotListDouble_t * destroy_operation_list_element( IotListDouble_t * pElt )
-{
-    if( pElt == NULL )
-    {
-        return NULL;
-    }
-
-    IotListDouble_t * pNext = pElt->pNext;
-    free( IotLink_Container( struct _mqttOperation, pElt, link ) );
-    return pNext;
-}
-
-void destroy_operation_list( IotListDouble_t * pList )
-{
-    if( pList == NULL )
-    {
-        return;
-    }
-
-    /* Assuming lists are of length at most 3 */
-
-    IotListDouble_t * pElt = pList->pNext;
-
-    if( pElt != pList )
-    {
-        pElt = destroy_operation_list_element( pElt );
-    }
-
-    if( pElt != pList )
-    {
-        pElt = destroy_operation_list_element( pElt );
-    }
-
-    if( pElt != pList )
-    {
-        pElt = destroy_operation_list_element( pElt );
-    }
-
-    pList->pNext = invalid_pointer();
-    pList->pPrevious = invalid_pointer();
-}
-
-/****************************************************************
-* Abstract the Notify method
-*
-* This abstraction assumes that the write set of Notify is just the
-* subscription and operation lists in the connection, and it models
-* Notify by simply havocing these lists.
-****************************************************************/
-
-void _IotMqtt_Notify( _mqttOperation_t * pOperation )
-{
-    destroy_subscription_list( &pOperation->pMqttConnection->subscriptionList );
-    destroy_operation_list( &pOperation->pMqttConnection->pendingProcessing );
-    destroy_operation_list( &pOperation->pMqttConnection->pendingResponse );
-}
-
-/****************************************************************
 * A predicate asserting that an operation on the pending list with no
 * retry has a job reference of 1.  This property is asserted by
 * _IotMqtt_FindOperation.
