+++ cbmc/proofs/IotMqtt_ReceiveCallback/Makefile
@@ -55,6 +55,11 @@ DEF += -DCBMC_STUB_REMOVEALLMATCHES=1
 
 OBJS += $(MQTT)/cbmc/proofs/$(ENTRY)/$(ENTRY)_harness.goto
 OBJS += $(MQTT)/cbmc/proofs/mqtt_state.goto
+OBJS += $(MQTT)/cbmc/stubs/_IotMqtt_Notify.goto
+OBJS += $(MQTT)/cbmc/stubs/IotListDouble_RemoveAll.goto
+OBJS += $(MQTT)/cbmc/stubs/IotListDouble_RemoveAllMatches.goto
+OBJS += $(MQTT)/cbmc/stubs/IotTaskPool_CreateJob.goto
+OBJS += $(MQTT)/cbmc/stubs/IotTaskPool_ScheduleDeferred.goto
 OBJS += $(MQTT)/libraries/standard/mqtt/src/iot_mqtt_api.goto
 OBJS += $(MQTT)/libraries/standard/mqtt/src/iot_mqtt_helper.goto
 OBJS += $(MQTT)/libraries/standard/mqtt/src/iot_mqtt_network.goto
