+++ cbmc/proofs/IotMqtt_UnsubscribeAsync/IotMqtt_UnsubscribeAsync_harness.c
@@ -35,87 +35,6 @@
 
 #include "mqtt_state.h"
 
-/****************************************************************/
-
-/**
- * We abstract these functions for performance reasons. In their original
- * implementation, CBMC ended up creating byte extracts for all possible
- * objects, due to the polymorphic nature of the linked list. We assume
- * that the functions are memory safe, and we free and havoc the list to
- * demonstrate that no subsequent code makes use of the values in the list.
- */
-
-typedef bool ( * MatchFunction_t )( const IotLink_t * const pOperationLink,
-                                    void * pCompare );
-typedef void ( * FreeElementFunction_t )( void * pData );
-
-void IotListDouble_RemoveAllMatches( const IotListDouble_t * const pList,
-                                     MatchFunction_t isMatch,
-                                     void * pMatch,
-                                     FreeElementFunction_t freeElement,
-                                     size_t linkOffset )
-{
-    free_IotMqttSubscriptionList( pList );
-    allocate_IotMqttSubscriptionList( pList, SUBSCRIPTION_COUNT_MAX - 1 );
-}
-
-/****************************************************************/
-
-/**
- * We constrain the return values of these functions because
- * they are checked by assertions in the MQTT code.
- */
-
-IotTaskPoolError_t IotTaskPool_CreateJob( IotTaskPoolRoutine_t userCallback,
-                                          void * pUserContext,
-                                          IotTaskPoolJobStorage_t * const pJobStorage,
-                                          IotTaskPoolJob_t * const pJob )
-{
-    assert( userCallback != NULL );
-    assert( pJobStorage != NULL );
-    assert( pJob != NULL );
-
-    /* _IotMqtt_ScheduleOperation asserts this */
-    return IOT_TASKPOOL_SUCCESS;
-}
-
-IotTaskPoolError_t IotTaskPool_ScheduleDeferred( IotTaskPool_t taskPool,
-                                                 IotTaskPoolJob_t job,
-                                                 uint32_t timeMs )
-{
-    IotTaskPoolError_t error;
-
-    /* _IotMqtt_ScheduleOperation asserts this */
-    __CPROVER_assume( error != IOT_TASKPOOL_BAD_PARAMETER &&
-                      error != IOT_TASKPOOL_ILLEGAL_OPERATION );
-    return error;
-}
-
-/****************************************************************/
-
-/**
- * _IotMqtt_NextPacketIdentifier calls Atomic_Add_u32, which receives
- * a volatile variable as input. Thus, CBMC will always consider that
- * Atomic_Add_u32 will operate over nondetermistic values and it raises
- * an unsigned integer overflow failure. However, developers have reported
- * that the use of this overflow is part of the function implementation.
- * In order to mirror _IotMqtt_NextPacketIdentifier behaviour and avoid
- * spurious alarms, we stub out this function to always
- * return a nondetermistic odd value.
- */
-
-uint16_t _IotMqtt_NextPacketIdentifier( void )
-{
-    uint16_t id;
-
-    /* Packet identifiers will follow the sequence 1,3,5...65535,1,3,5... */
-    __CPROVER_assume( id & 0x0001 == 1 );
-
-    return id;
-}
-
-/****************************************************************/
-
 void harness()
 {
     size_t subscriptionCount;
@@ -145,8 +64,6 @@ void harness()
     /* output */
     IotMqttOperation_t pUnsubscribeOperation =
         allocate_IotMqttOperation( NULL, mqttConnection );
-
-
 
     IotMqtt_UnsubscribeAsync( mqttConnection,
                               subscriptionList,
