+++ cbmc/proofs/IotMqtt_UnsubscribeSync/IotMqtt_UnsubscribeSync_harness.c
@@ -35,98 +35,6 @@
 
 #include "mqtt_state.h"
 
-/****************************************************************/
-
-/**
- * We abstract this function for performance reasons. In its original
- * implementation, CBMC ended up creating byte extracts for all possible
- * objects, due to the polymorphic nature of the linked list. We assume
- * that the function is memory safe, and we free and havoc the list to
- * demonstrate that no subsequent code makes use of the values in the list.
- */
-
-typedef bool ( * MatchFunction_t )( const IotLink_t * const pOperationLink,
-                                    void * pCompare );
-typedef void ( * FreeElementFunction_t )( void * pData );
-
-void IotListDouble_RemoveAllMatches( const IotListDouble_t * const pList,
-                                     MatchFunction_t isMatch,
-                                     void * pMatch,
-                                     FreeElementFunction_t freeElement,
-                                     size_t linkOffset )
-{
-    free_IotMqttSubscriptionList( pList );
-    allocate_IotMqttSubscriptionList( pList, SUBSCRIPTION_COUNT_MAX - 1 );
-}
-
-/****************************************************************/
-
-/**
- * We constrain the return values of these functions because they
- * are checked by assertions in the MQTT code.
- */
-
-IotTaskPoolError_t IotTaskPool_TryCancel( IotTaskPool_t taskPool,
-                                          IotTaskPoolJob_t job,
-                                          IotTaskPoolJobStatus_t * const pStatus )
-{
-    assert( taskPool == NULL || job == NULL );
-    return IOT_TASKPOOL_BAD_PARAMETER;
-}
-
-/****************************************************************/
-
-/**
- * _IotMqtt_NextPacketIdentifier calls Atomic_Add_u32, which receives
- * a volatile variable as input. Thus, CBMC will always consider that
- * Atomic_Add_u32 will operate over nondetermistic values and it raises
- * an unsigned integer overflow failure. However, developers have reported
- * that the use of this overflow is part of the function implementation.
- * In order to mirror _IotMqtt_NextPacketIdentifier behaviour and avoid
- * spurious alarms, we stub out this function to always
- * return a nondetermistic odd value.
- */
-uint16_t _IotMqtt_NextPacketIdentifier( void )
-{
-    uint16_t id;
-
-    /* Packet identifiers will follow the sequence 1,3,5...65535,1,3,5... */
-    __CPROVER_assume( id & 0x0001 == 1 );
-
-    return id;
-}
-
-/****************************************************************/
-
-/**
- * We are abstracting the semaphores because we are doing sequential proof.
- * But the semaphore API assures us that TimedWait called after Post will
- * never fail. Our abstraction of the semaphores models this behavior.
- */
-static unsigned int flagSemaphore;
-
-void IotSemaphore_Post( IotSemaphore_t * pSemaphore )
-{
-    assert( pSemaphore != NULL );
-    flagSemaphore++;
-}
-
-bool IotSemaphore_TimedWait( IotSemaphore_t * pSemaphore,
-                             uint32_t timeoutMs )
-{
-    assert( pSemaphore != NULL );
-
-    if( flagSemaphore > 0 )
-    {
-        flagSemaphore--;
-        return true;
-    }
-
-    return false;
-}
-
-/****************************************************************/
-
 void harness()
 {
     size_t subscriptionCount;
@@ -149,9 +57,6 @@ void harness()
 
     uint32_t flags;
     uint32_t timeoutMs;
-
-    /* initialize semaphore flag */
-    flagSemaphore = 0;
 
     IotMqtt_UnsubscribeSync( mqttConnection,
                              subscriptionList,
